% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr_overrides.R
\name{add_dataframe}
\alias{add_dataframe}
\title{Adds a dataframe to}
\usage{
add_dataframe(table, filename, group)
}
\arguments{
\item{table}{a metadata section from the CLDF metadata.}

\item{filename}{the filename.}

\item{group}{a grouping from the metadata.}
}
\value{
A dataframe
}
\description{
Adds a dataframe to
}
