% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{citation}
\alias{citation}
\title{Extracts a single table from a CLDF dataset.}
\usage{
citation(object)
}
\arguments{
\item{object}{the CLDF dataset}
}
\value{
invisible(string)
}
\description{
Extracts a single table from a CLDF dataset.
}
\examples{
citation(cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf")))
}
