% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr_overrides.R
\name{coalesce_truth}
\alias{coalesce_truth}
\title{Coalesce value to truthiness}
\usage{
coalesce_truth(x)
}
\arguments{
\item{x}{logical, \code{NA} or \code{NULL}}
}
\value{
\code{FALSE} if x is anything but \code{TRUE}
}
\description{
Determine whether the input is true, with missing values being interpreted as false.
}
