% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr_overrides.R
\name{default_schema}
\alias{default_schema}
\title{Create a default table schema given a csv file and dialect}
\usage{
default_schema(filename, dialect = default_dialect)
}
\arguments{
\item{filename}{a csv file}

\item{dialect}{specification of the csv's dialect (default: \code{default_dialect})}
}
\value{
a table schema
}
\description{
If neither the table nor the group have a \code{tableSchema} annotation,
then this default schema will used.
}
