% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Downloads and unpacks a dataset into the user cache dir}
\usage{
download(url, cache_dir = tools::R_user_dir("rcldf", which = "cache"))
}
\arguments{
\item{url}{a string}

\item{cache_dir}{a path to the desired cache directory}
}
\value{
a string containing the download path
}
\description{
Downloads and unpacks a dataset into the user cache dir
}
\examples{
o <- download("https://github.com/lexibank/panobodyparts/archive/refs/tags/v1.0.zip")
}
