% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_citation.R
\name{get_citation}
\alias{get_citation}
\title{Prints and returns the Citation for the given CLDF dataset}
\usage{
get_citation(obj)
}
\arguments{
\item{obj}{A CLDF object}
}
\value{
An (invisible) string containing the citation text
}
\description{
Prints and returns the Citation for the given CLDF dataset
}
\examples{
cldfobj <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
get_citation(cldfobj)
}
