% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details}
\alias{get_details}
\title{Returns a dataframe of with details on the CLDF dataset in \code{path}.}
\usage{
get_details(path)
}
\arguments{
\item{path}{the path to resolve}
}
\value{
A dataframe.
}
\description{
Returns a dataframe of with details on the CLDF dataset in \code{path}.
}
