% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_dir_size}
\alias{get_dir_size}
\title{Returns the filesize in bytes of a directory.}
\usage{
get_dir_size(path)
}
\arguments{
\item{path}{a directory to size}
}
\value{
A numeric of the file size in bytes
}
\description{
Returns the filesize in bytes of a directory.
}
