% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filename.R
\name{get_filename}
\alias{get_filename}
\title{Get a filename from url value in metadata (handles .zip files)}
\usage{
get_filename(base_dir, url)
}
\arguments{
\item{base_dir}{the base_dir}

\item{url}{the url statement}
}
\value{
A string
}
\description{
Get a filename from url value in metadata (handles .zip files)
}
