% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_glottolog.R
\name{get_from_zenodo}
\alias{get_from_zenodo}
\title{Downloads and installs a CLDF dataset from a Zenodo endpoint}
\usage{
get_from_zenodo(zid, load_bib = FALSE, cache_dir = NULL)
}
\arguments{
\item{zid}{Zenodo endpoint \code{conceptid}}

\item{load_bib}{load sources (TRUE/FALSE, default FALSE)}

\item{cache_dir}{A cache_dir to use. If NULL it will use get_cache_dir}
}
\value{
A \code{cldf} object
}
\description{
Downloads and installs a CLDF dataset from a Zenodo endpoint
}
