% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{get_separators}
\alias{get_separators}
\title{Identifies the separator characters specified by the CLDF metadata.}
\usage{
get_separators(metadata)
}
\arguments{
\item{metadata}{\itemize{
\item a CLDF metadata.
}}
}
\value{
A dataframe with three columns (name, separator, url).
}
\description{
Identifies the separator characters specified by the CLDF metadata.
}
