% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_from.R
\name{get_table_from}
\alias{get_table_from}
\title{Extracts a single table from a CLDF dataset.}
\usage{
get_table_from(table, mdpath)
}
\arguments{
\item{table}{a CLDF table type}

\item{mdpath}{a path to a CLDF file}
}
\value{
a dataframe
}
\description{
Extracts a single table from a CLDF dataset.
}
\examples{
df <- get_table_from("LanguageTable", "my/cldf")
}
