% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_github.R
\name{is_github}
\alias{is_github}
\title{Returns TRUE if \code{url} looks like a github URL}
\usage{
is_github(url)
}
\arguments{
\item{url}{A string}
}
\value{
A boolean TRUE/FALSE
}
\description{
Returns TRUE if \code{url} looks like a github URL
}
\examples{
is_github('https://github.com/SimonGreenhill/rcldf/')
}
