% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url.R
\name{is_url}
\alias{is_url}
\title{Returns TRUE if \code{url} looks like a URL}
\usage{
is_url(url)
}
\arguments{
\item{url}{A string}
}
\value{
A boolean TRUE/FALSE
}
\description{
Returns TRUE if \code{url} looks like a URL
}
\examples{
is_url('http://simon.net.nz')
}
