% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{list_cache_files}
\alias{list_cache_files}
\title{Returns a dataframe of directories in the cache dir}
\usage{
list_cache_files()
}
\value{
A dataframe of the directories
}
\description{
Returns a dataframe of directories in the cache dir
}
