% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_glottolog.R
\name{load_concepticon}
\alias{load_concepticon}
\title{Returns a CLDF dataset object of the latest concepticon version.}
\usage{
load_concepticon(path, load_bib = FALSE, cache_dir = NULL)
}
\arguments{
\item{path}{the path to resolve}

\item{load_bib}{load sources (TRUE/FALSE, default FALSE)}

\item{cache_dir}{A cache_dir to use. If NULL it will use get_cache_dir}
}
\value{
A \code{cldf} object
}
\description{
Returns a CLDF dataset object of the latest concepticon version.
}
