% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{make_cache_key}
\alias{make_cache_key}
\title{Returns the cachekey for the given path.}
\usage{
make_cache_key(path)
}
\arguments{
\item{path}{a path to generate the cachekey for.}
}
\value{
A string.
}
\description{
Returns the cachekey for the given path.
}
