% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr_overrides.R
\name{override_defaults}
\alias{override_defaults}
\title{Override defaults}
\usage{
override_defaults(...)
}
\arguments{
\item{...}{any number of lists with configuration values}
}
\value{
a list with the values from the first list replacing those in the second and so on
}
\description{
Merges two lists applying \code{override} values on top of the \code{default} values.
}
\keyword{internal}
