% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{print.cldf_schema}
\alias{print.cldf_schema}
\title{Prints a CLDF schema}
\usage{
\method{print}{cldf_schema}(x, ...)
}
\arguments{
\item{x}{the CLDF dataset}

\item{...}{Arguments to be passed to or from other methods. Currently not used.}
}
\value{
No return value, called for side effects.
}
\description{
Prints a CLDF schema
}
\examples{
cldfobj <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
print(schema(cldfobj))
}
