% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rcldf-package}
\alias{rcldf}
\alias{rcldf-package}
\title{rcldf: Read Linguistic Data in the Cross Linguistic Data Format (CLDF)}
\description{
Cross-Linguistic Data Format (CLDF) is a framework for storing cross-linguistic data, ensuring compatibility and ease of data exchange between different linguistic datasets see Forkel et al. (2018) \doi{10.1038/sdata.2018.205}. The 'rcldf' package is designed to facilitate the manipulation and analysis of these datasets by simplifying the loading, querying, and visualisation of CLDF datasets making it easier to conduct comparative linguistic analyses, manage language data, and apply statistical methods directly within R.
}
\details{
rcldf is a library for R to read Cross-Linguistic Data files (CLDF)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SimonGreenhill/rcldf}
  \item Report bugs at \url{https://github.com/SimonGreenhill/rcldf/issues}
}

}
\author{
\strong{Maintainer}: Simon J. Greenhill \email{simon@simon.net.nz}

}
\keyword{internal}
