% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel}
\alias{relabel}
\title{Relabels a column in a dataset for merging.}
\usage{
relabel(column, table)
}
\arguments{
\item{column}{the tablename.}

\item{table}{the tablename.}
}
\value{
A string of "column.table"
}
\description{
Relabels a column in a dataset for merging.
}
