% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_path.R
\name{resolve_path}
\alias{resolve_path}
\title{Helper function to resolve the path (e.g. directory or md.json file)}
\usage{
resolve_path(path)
}
\arguments{
\item{path}{the path to resolve}
}
\value{
A string containing the path to the metadata.json file
}
\description{
Helper function to resolve the path (e.g. directory or md.json file)
}
