% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{separate}
\alias{separate}
\title{Expands all values with separators.}
\usage{
separate(cldfobj, separators = NULL)
}
\arguments{
\item{cldfobj}{a CLDF Object}

\item{separators}{a dataframe of separator values to replace (default=NULL).}
}
\value{
A \code{cldf} object
}
\description{
Note that this is run by default on loading a dataset with cldf()
}
\examples{
cldfobj <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
cldfobj <- separate(cldfobj)
}
