% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{set_cache_dir}
\alias{set_cache_dir}
\title{Sets the cache dir for the current session.}
\usage{
set_cache_dir(cache_dir = NA)
}
\arguments{
\item{cache_dir}{a directory to use}
}
\value{
NULL. Sets an environment value.
}
\description{
Sets the cache dir for the current session.
}
