% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cldf.R
\name{summary.cldf}
\alias{summary.cldf}
\title{Summarises the CLDF file}
\usage{
\method{summary}{cldf}(object, ...)
}
\arguments{
\item{object}{the CLDF dataset}

\item{...}{Arguments to be passed to or from other methods. Currently not used.}
}
\value{
None
}
\description{
Summarises the CLDF file
}
\examples{
cldfobj <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
summary(cldfobj)
}
