% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_cldf.R
\name{update_table}
\alias{update_table}
\title{Updates a table \code{tbl} based on expression \code{e}.}
\usage{
update_table(e, tbl)
}
\arguments{
\item{e}{the expression.}

\item{tbl}{the table.}
}
\value{
A filtered tables.
}
\description{
Helper function to filter a table  based on a logical expression.
}
