% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcldf.R
\name{cldf}
\alias{cldf}
\alias{read_cldf}
\title{Reads a Cross-Linguistic Data Format dataset into an object.}
\usage{
cldf(mdpath)

read_cldf(path)
}
\arguments{
\item{mdpath}{the path to the directory or metadata.json file.}
}
\value{
A `cldf` object
}
\description{
Reads a Cross-Linguistic Data Format dataset into an object.
}
