% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{Returns a table of datasets available in cldf_meta}
\usage{
datasets()
}
\value{
A dataframe of available dataset.
}
\description{
Returns a table of datasets available in cldf_meta
}
