% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{list_cache_files}
\alias{list_cache_files}
\title{Returns a dataframe of directories in the cache dir}
\usage{
list_cache_files(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{the cache directory to use. If NULL then R_user_dir will be used.}
}
\value{
A dataframe of the directories
}
\description{
Returns a dataframe of directories in the cache dir
}
