% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_glottolog.R
\name{load_glottolog}
\alias{load_glottolog}
\title{Returns a CLDF dataset object of the latest glottolog version.}
\usage{
load_glottolog(path, cache_dir = NULL)
}
\arguments{
\item{path}{the path to resolve}

\item{cache_dir}{A cache_dir to use. If NULL it will use get_cache_dir}
}
\value{
A \code{cldf} object
}
\description{
Returns a CLDF dataset object of the latest glottolog version.
}
