% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cldf.R
\name{print.cldf}
\alias{print.cldf}
\title{Summarises the CLDF file}
\usage{
\method{print}{cldf}(x, ...)
}
\arguments{
\item{x}{the CLDF dataset}

\item{...}{Arguments to be passed to or from other methods. Currently not used.}
}
\value{
No return value, called for side effects.
}
\description{
Summarises the CLDF file
}
\examples{
cldfobj <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
print(cldfobj)
}
