% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rcldf-package}
\alias{rcldf}
\alias{rcldf-package}
\title{rcldf: rcldf - Read Linguistic Data In The Cross Linguistic Data Format (CLDF)}
\description{
Reads Cross-Linguistic Data Files (CLDF) into R.
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()].
Otherwise refer to the vignettes to see how to format the documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SimonGreenhill/rcldf}
  \item Report bugs at \url{https://github.com/SimonGreenhill/rcldf/issues}
}

}
\keyword{internal}
