% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rcldf-package}
\alias{rcldf}
\alias{rcldf-package}
\title{rcldf: Read Linguistic Data in the Cross Linguistic Data Format (CLDF)}
\description{
Reads Cross-Linguistic Data Files (CLDF) into R.
}
\details{
rcldf is a library for R to read Cross-Linguistic Data files (CLDF)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SimonGreenhill/rcldf}
  \item Report bugs at \url{https://github.com/SimonGreenhill/rcldf/issues}
}

}
\author{
\strong{Maintainer}: Simon J. Greenhill \email{simon@simon.net.nz}

}
\keyword{internal}
