% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bib.R
\name{read_bib}
\alias{read_bib}
\title{Reads a BibTeX file into a dataframe}
\usage{
read_bib(object)
}
\arguments{
\item{object}{A CLDF object}
}
\value{
A tibble dataframe
}
\description{
Reads a BibTeX file into a dataframe
}
