% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_path.R
\name{resolve_path}
\alias{resolve_path}
\title{Helper function to resolve the path (e.g. directory or md.json file)}
\usage{
resolve_path(path, cache_dir = NA)
}
\arguments{
\item{path}{the path to resolve}

\item{cache_dir}{a directory to cache downloaded files to}
}
\value{
A list of two items:
\code{path}  - string containing the path to the metadata.json file
\code{metadata} - a csvwr metadata object
}
\description{
Helper function to resolve the path (e.g. directory or md.json file)
}
